모바일 내점심줄 개발완료

+++ src/hooks/useFilter.ts
@@ -12,6 +12,7 @@ function useFilter() {
   const [selectedPlaceList, setSelectedPlaceList] = useState<
     AccountPlaceInfo[]
   >();
+  const [insertPlaceSuccess, setInsertPlaceSuccess] = useState(false);
 
   const asyncInsertSelectedPlace = useCallback(
     async (requestBody: {
@@ -23,11 +24,14 @@ function useFilter() {
         const { data } = await insertSelectedPlace(requestBody);
         if (data.resultCode === 200) {
           window.alert('자주 이용하는 음식점이 저장되었습니다.');
+          setInsertPlaceSuccess(true);
         } else {
           window.alert(COMMON_MESSAGE.FILTER_MESSAGE.SAVE_FAIL);
+          setInsertPlaceSuccess(false);
         }
       } catch (e) {
         window.alert(COMMON_MESSAGE.FILTER_MESSAGE.SAVE_FAIL);
+        setInsertPlaceSuccess(false);
       }
     },
     [],
@@ -61,8 +65,11 @@ function useFilter() {
       const { data } = await getSelectedPlace(id);
       if(data.resultCode === 200) {
         setSelectedPlaceList(data.data);
+      } else {
+        setSelectedPlaceList(undefined);
       }
     } catch (e) {
+      setSelectedPlaceList(undefined);
       window.alert(COMMON_MESSAGE.FILTER_MESSAGE.DATA_LOAD_FAIL);
     }
   }, []);
@@ -70,11 +77,14 @@ function useFilter() {
   const asyncDeleteSelectedPlace = useCallback(
     async (requestBody: { id: number; place_id: number }) => {
       try {
-        const { data } = await deleteSelectedPlace(requestBody);
-        if (data.resultCode === 200) {
-          window.alert('삭제되었습니다.');
-        } else {
-          window.alert(COMMON_MESSAGE.FILTER_MESSAGE.SAVE_FAIL);
+        if(window.confirm('삭제하시겠습니까?')) {
+          const { data } = await deleteSelectedPlace(requestBody);
+          if (data.resultCode === 200) {
+            window.alert('삭제되었습니다.');
+            asyncGetSelectedPlace(requestBody.id);
+          } else {
+            window.alert(COMMON_MESSAGE.FILTER_MESSAGE.SAVE_FAIL);
+          }
         }
       } catch (e) {
         window.alert(COMMON_MESSAGE.FILTER_MESSAGE.SAVE_FAIL);
@@ -85,6 +95,7 @@ function useFilter() {
 
   return {
     selectedPlaceList,
+    insertPlaceSuccess,
     asyncInsertSelectedPlace,
     asyncUpdateFilter,
     asyncGetSelectedPlace,
